feat: merge develop (0.8.2) into master

+++ component/keyple-calypso/src/main/command/sam/builder/security/SamReadCeilingsCmdBuild.h
@@ -56,7 +56,7 @@ public:
      *
      */
     SamReadCeilingsCmdBuild(SamRevision& revision,
-                            CeilingsOperationType operationType, int index);
+                            CeilingsOperationType operationType, uint8_t index);
 
 private:
     /**
